requirements.txt
beautifulsoup4==4.12.2
numpy==1.26.2
pandas==2.2.0
openpyxl==3.1.2
pydantic==2.4.2
python_dateutil==2.8.2
PyYAML==6.0.1
Requests==2.31.0
typer==0.9.0
typing_extensions==4.8.0
fdb  # Firebird database adapter
